refactor: rename `RoomListView` as `RoomListPanel`
