+++ res/css/views/rooms/RoomListPanel/_RoomListHeaderView.pcss
