+++ res/css/views/rooms/RoomListPanel/_RoomListPanel.pcss
@@ -5,4 +5,8 @@
  * Please see LICENSE files in the repository root for full details.
  */
 
-export { RoomListView } from "./RoomListView";
+.mx_RoomListPanel {
+    background-color: var(--cpd-color-bg-canvas-default);
+    height: 100%;
+    border-right: 1px solid var(--cpd-color-bg-subtle-primary);
+}
