+++ res/css/views/rooms/RoomListPanel/_RoomListSearch.pcss
